rename buttons to buttonStyle and typeScale and textStyle
